+++ src/client/commands/mod.rs
@@ -1,4 +1,5 @@
 mod deploy;
+mod down;
 mod logs;
 mod ps;
 mod restart;
@@ -6,6 +7,7 @@ mod run;
 mod stop;
 
 pub use self::deploy::*;
+pub use self::down::*;
 pub use self::logs::*;
 pub use self::ps::*;
 pub use self::restart::*;